+++ MyGUIEngine/src/MyGUI_RotatingSkin.cpp
@@ -290,7 +290,7 @@ namespace MyGUI
 		// now we are going to calculate verticies of resulting figure
 
 		// no parent - no cropping
-		int size = RECT_VERTICIES_COUNT;
+		size_t size = RECT_VERTICIES_COUNT;
 		if (nullptr == mCroppedParent->getCroppedParent())
 		{
 			for (int i = 0; i < RECT_VERTICIES_COUNT; ++i)
@@ -324,7 +324,7 @@ namespace MyGUI
 			// calculate texture coordinates
 			FloatPoint v0 = baseVerticiesUV[1] - baseVerticiesUV[0];
 			FloatPoint v1 = baseVerticiesUV[3] - baseVerticiesUV[0];
-			for (int i = 0; i < GEOMETRY_VERTICIES_TOTAL_COUNT; ++i)
+			for (size_t i = 0; i < GEOMETRY_VERTICIES_TOTAL_COUNT; ++i)
 			{
 				if (i < size)
 				{
@@ -345,7 +345,7 @@ namespace MyGUI
 		float vertex_left_base = ((info.pixScaleX * (float)(mCroppedParent->getAbsoluteLeft()) + info.hOffset) * 2) - 1;
 		float vertex_top_base = -(((info.pixScaleY * (float)(mCroppedParent->getAbsoluteTop()) + info.vOffset) * 2) - 1);
 
-		for (int i = 0; i < GEOMETRY_VERTICIES_TOTAL_COUNT; ++i)
+		for (size_t i = 0; i < GEOMETRY_VERTICIES_TOTAL_COUNT; ++i)
 		{
 			if (i < size)
 			{
